11104 - move createFatLibrary to BundleHelper
